simple initiative tracker - heavily in progress
It is incredibly evident that this is not finished haha.
I started working on it today and does not do what I hope it will, yet.
Mainly, I was toying around with the idea before I even had a design of some sort.
I'll take some time tomorrow to create a  proper design. Then I will implement.